# lib/api-zod/src/generated/types/createTaskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTaskInputPriority } from './createTaskInputPriority';
import type { CreateTaskInputStatus } from './createTaskInputStatus';

export interface CreateTaskInput {
  title: string;
  description?: string | null;
  status?: CreateTaskInputStatus;
  priority?: CreateTaskInputPriority;
  dueDate?: Date | null;
  projectId: number;
  assignedToId?: number | null;
}