# lib/api-zod/src/generated/types/updateTaskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskInputPriority } from './updateTaskInputPriority';
import type { UpdateTaskInputStatus } from './updateTaskInputStatus';

export interface UpdateTaskInput {
  title?: string;
  description?: string | null;
  status?: UpdateTaskInputStatus;
  priority?: UpdateTaskInputPriority;
  dueDate?: Date | null;
  assignedToId?: number | null;
}